requirements.txt
argparse
tqdm
pycryptodome
pillow
numpy
prettytable
opencv-python
pandas
wget
requests
asyncio
prettyprint